Repositorio JPA

+++ src/main/java/br/com/fiap/api_rest/controller/ClienteController.java
@@ -5,6 +5,11 @@ import br.com.fiap.api_rest.dto.ClienteResponse;
 import br.com.fiap.api_rest.model.Cliente;
 import br.com.fiap.api_rest.repository.ClienteRepository;
 import br.com.fiap.api_rest.service.ClienteService;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.media.Schema;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,12 +21,11 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.List;
 import java.util.Optional;
 
 @RestController
 @RequestMapping(value = "/clientes", produces = {"application/json"})
-@Tag(name="api-clientes")
+@Tag(name = "api-clientes")
 public class ClienteController {
     @Autowired
     ClienteRepository clienteRepository;
@@ -31,12 +35,27 @@ public class ClienteController {
     // Create, Read, Update, Delete - CRUD
     // Post, Get, Put, Delete - Verbos HTTP correspondentes
 
+    @Operation(summary = "Cria um novo Cliente")
+    @ApiResponses(value = {
+            @ApiResponse(
+                    responseCode = "201",
+                    description = "Cliente Cadastrado com Sucesso!",
+                    content = {@Content(mediaType = "application/json",
+                            schema = @Schema(implementation = Cliente.class))}),
+
+            @ApiResponse(
+                    responseCode = "400",
+                    description = "Atributos Inválidos.",
+                    content = {@Content(schema = @Schema())})
+
+    })
     @PostMapping
     public ResponseEntity<Cliente> createCliente(@Valid @RequestBody ClienteRequest cliente) {
         Cliente clienteSalvo = clienteRepository.save(clienteService.requestToCliente(cliente));
         return new ResponseEntity<>(clienteSalvo, HttpStatus.CREATED);
     }
 
+    @Operation(summary = "Retorna uma Lista de Clientes")
     @GetMapping
     public ResponseEntity<Page<ClienteResponse>> readClientes(@RequestParam(defaultValue = "0") Integer page) {
         Pageable pageable = PageRequest.of(page, 2, Sort.by("categoria").ascending().and(Sort.by("nome").ascending()));
@@ -45,6 +64,21 @@ public class ClienteController {
 
     // PathVariable = parâmetro diretamente na URL, ex: /clientes/1
     // RequestParam = parâmetro como query, ex: /clientes/?id=1
+
+    @Operation(summary = "Retorna Um Cliente por ID")
+    @ApiResponses(value = {
+            @ApiResponse(
+                    responseCode = "200",
+                    description = "Cliente Encontrado com Sucesso!",
+                    content = {@Content(mediaType = "application/json",
+                            schema = @Schema(implementation = ClienteResponse.class))}),
+
+            @ApiResponse(
+                    responseCode = "404",
+                    description = "Cliente Não Encontrado.",
+                    content = {@Content( schema = @Schema())})
+
+    })
     @GetMapping("/{id}")
     public ResponseEntity<ClienteResponse> readCliente(@PathVariable Long id) {
         Optional<Cliente> cliente = clienteRepository.findById(id);
@@ -54,6 +88,20 @@ public class ClienteController {
         return new ResponseEntity<>(clienteService.clienteToResponse(cliente.get(), false), HttpStatus.OK);
     }
 
+    @Operation(summary = "Atualiza um Cliente existente")
+    @ApiResponses(value = {
+            @ApiResponse(
+                    responseCode = "200",
+                    description = "Cliente Encontrado e Atualizado com Sucesso!",
+                    content = {@Content(mediaType = "application/json",
+                            schema = @Schema(implementation = Cliente.class))}),
+
+            @ApiResponse(
+                    responseCode = "404",
+                    description = "Cliente Não Encontrado.",
+                    content = {@Content( schema = @Schema())})
+
+    })
     @PutMapping("/{id}")
     public ResponseEntity<Cliente> updateCliente(@PathVariable Long id,
                                                  @RequestBody Cliente cliente) {
@@ -66,8 +114,26 @@ public class ClienteController {
         return new ResponseEntity<>(clienteAtualizado, HttpStatus.CREATED);
     }
 
+    @Operation(summary = "Deleta um Cliente existente")
+    @ApiResponses(value = {
+            @ApiResponse(
+                    responseCode = "200",
+                    description = "Cliente Deletado com Sucesso!",
+                    content = {@Content(mediaType = "application/json",
+                            schema = @Schema(implementation = Cliente.class))}),
+
+            @ApiResponse(
+                    responseCode = "404",
+                    description = "Cliente Não Encontrado.",
+                    content = {@Content( schema = @Schema())})
+
+    })
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deleteCliente(@PathVariable Long id) {
+        Optional<Cliente> clienteExistente = clienteRepository.findById(id);
+        if (clienteExistente.isEmpty()) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
         clienteRepository.deleteById(id);
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
